Refactor(#65): album listener 패키지 분리
