+++ src/main/java/com/widyu/domain/fcm/event/album/dto/AlbumCreatedEvent.java
@@ -1,4 +1,4 @@
-package com.widyu.domain.fcm.event.album;
+package com.widyu.domain.fcm.event.album.dto;
 
 public record AlbumCreatedEvent(Long albumId, Long authorId) {
 }
